Splitting Writer class into 3 seperate classes.
This change splits Writer into GlobalOrderWriter, UnorderedWriter and
OrderedWriter. The code changes were minimally done to query to create
the appropriate class depending on the write type, everything else is
just code movement.

---
TYPE: IMPROVEMENT
DESC: Splitting Writer class into 3 seperate classes.
